[sink] Change voice call volume control implementation.
Previously when voice volume control was enabled, droid sink switched to
HW volume control, and HW volume control changes affected voice call
volume. The problem with this is, regular audio mixed to the voice call
do not follow the voice call volume, thus non-call audio mixed to the
call had correct volume level only if sink's volume was 0dB.

Change the implementation so that when voice call volume control is
enabled, track for specific sink-input with key:value (defaults to
"media.role":"phone") and when sink-input with correct proplist
key-value pair connects to the sink, use the sink-input's absolute volume
for controlling the voice call volume.