[sink] Sleep after stream writing blocks. Fixes MER#1212
After HAL buffer is full and stream writing starts to block close to the
buffer length, sleep for one buffer length time. When io thread is
sleeping the thread can be waken up for different reasons, and if we
constantly block in the write function io thread is unable to react
immediately.