[glue] New module for forwarding AudioFlinger calls. MER#1377
Simple module that forwards calls from afglue library to audio HAL hw
module. As with droid modules the HAL instance is not running inside
AudioFlinger service for example RIL doesn't have any means to call into
the HAL. This new module forwards the calls so that AudioFlinger users
are happy.