[util] Add new functions for stream operations. MER#1419
Add functions for opening output and input streams, setting output and
input stream route, and setting parameters to streams and hw module.

The greatest improvement this brings is output stream routing handling
in one place. This way if/when there are multiple output streams routing
changes done to primary output are forwarded to "secondary" or "slave"
output streams.