[source] Only refresh routing if forced. JB#34890
Unless the routing is forced do not refresh routing. With output routing
it is necessary to refresh the routing when changing audio mode, but
input routing doesn't have similar requirement. Also on some devices
setting identical input routing blocks thread for a while. Due to this
disable input route refreshing for most cases.