[source] Close input stream by default. Fixes JB#38419
When changing source port, by default we will now close the input
stream, set the new parameters, and reopen the stream. At the same time
it is possible that the channel count or buffer size of the input stream
needs to change (mainly for voicecall recording case, where we need to
use mono channel map for best compatibility across different devices).
It is also possible to revert to older behavior where input stream is
always around, using quirks=-close_input

We also now start droid source always in suspended state. We still need
to open the input stream for a while to get stream buffer size and
sample rate.