card: read headphone availability from input device
Recent Android devices start to provide headphone availability via input
device instead of h2w switch. This renders droid-extcon useless.

This commit introduce droid-extevdev, a simple code that will read
headphone availability from the input device, using libevdev as an
abstraction layer. This means the code now depends on libevdev (but it
can be made optional later on if needed).

This should make headphone availability works on newer Android devices
without having to resort to h2w kernel driver.

PATCH: extevdev: make the combo port always available

Ringtone & alarm uses the combo port to play to both the headphone and
the speaker. But because we distinguish between headphone and headset,
the port won't be available if we plug a headset with a mic.

Turns out, droid-extcon doesn't do the availability thing on this port,
so the port is always available. So, we should do the same.

[card] read headphone availability from input device.

Merged from UBports.