source: Add a workaround for fm-radio loopback.
Refactor reconfiguring part a bit and add a workaround for fm-radio
loopback. As the loopback module is instantiated without defined sink
or source, it uses as default really silly values. But since our source
is reconfiguring itself to whatever is requested things get a bit
hairy. To ensure good values for the source in case of loopback module
connection use metrics from our primary output stream.

[source] Add a workaround for fm-radio loopback. JB#48080