common: Add quirk audio_cal_wait.
If there is a device which runs threaded audio calibration after hw
module open we need to wait for a while after opening the first output
stream.

Also when the calibration is already done and we are starting we again
need to wait for a while for the calibration loading to finish.

[common] Add quirk audio_cal_wait. JB#52328